requirements.txt
openai
python-dotenv
transformers
torch
PyPDF2
supabase
boto3
google-cloud-vision
python-bidi==0.4.2
llama-index
faiss-cpu
fpdf2
python-bidi
faiss
numpy==1.24.3
llama-index-vector-stores-faiss
llama-index-vector-stores-postgres